pyproject.toml
[tool.poetry]
name = "emaildownloader"
version = "0.1.0"
description = ""
authors = ["ia rigby <marvinzem@gmail.com>"]
packages = [
    {include = "email_attachment_downloader"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.147.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pikepdf = "^9.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
